fix: ensure map is fully saved before quitting on Windows
Windows users were reporting 0-byte map files when saving upon exit.
This was caused by a race condition where MapDestination::alloc()
(via FileSaver::open()) truncated the file in the main thread, but
the application terminated before the background thread could write
the data.

This patch:
1. Re-introduces waitForAsync() to block the main thread until
   background saving is complete during critical transitions.
2. Implements sig_asyncTaskFinished to synchronize between the
   AsyncTask and the main window's event loop.
3. Enables atomic saving on all platforms (including Windows) by
   using a .tmp suffix and robust replacement APIs (ReplaceFileW
   on Windows, rename on POSIX), ensuring the original file is
   preserved until the new data is securely written.
4. Implements io::fsync() for Windows using FlushFileBuffers to
   ensure data integrity.
5. Improves support for long filenames on Windows by utilizing
   Unicode APIs.
6. Simplifies FileSaver code by removing now-redundant platform
   toggles for atomic saving.
7. Refactors platform-specific rename logic into io::rename.
