+++ src/mapstorage/filesaver.cpp
@@ -12,10 +12,6 @@
 #include <stdexcept>
 #include <tuple>
 
-#ifdef Q_OS_WIN
-#include <windows.h>
-#endif
-
 #include <QIODevice>
 
 static const char *const TMP_FILE_SUFFIX = ".tmp";
@@ -29,28 +25,7 @@ static void remove_tmp_suffix(const QString &filename) CAN_THROW
     const QString from = filename + TMP_FILE_SUFFIX;
     const QString to = filename;
 
-#ifdef Q_OS_WIN
-    const std::wstring fromW = from.toStdWString();
-    const std::wstring toW = to.toStdWString();
-    if (::ReplaceFileW(toW.c_str(), fromW.c_str(), nullptr, REPLACEFILE_IGNORE_MERGE_ERRORS, nullptr, nullptr)
-        == 0) {
-        const auto err = ::GetLastError();
-        if (err == ERROR_FILE_NOT_FOUND) {
-            if (::MoveFileExW(fromW.c_str(), toW.c_str(), MOVEFILE_REPLACE_EXISTING | MOVEFILE_COPY_ALLOWED)
-                == 0) {
-                throw io::IOException::withErrorNumber(static_cast<int>(::GetLastError()));
-            }
-        } else {
-            throw io::IOException::withErrorNumber(static_cast<int>(err));
-        }
-    }
-#else
-    const auto fromEncoded = QFile::encodeName(from);
-    const auto toEncoded = QFile::encodeName(to);
-    if (::rename(fromEncoded.data(), toEncoded.data()) == -1) {
-        throw io::IOException::withCurrentErrno();
-    }
-#endif
+    io::rename(from, to);
 }
 
 FileSaver::~FileSaver()
